refactor from instructor views to analytics module
